Implemented basic logic to set picture details.
Still very experimental and weak, but it already allows to set some basic
properties for every picture. Needs a lot of polishing and work to avoid
error cases (nulls, empty fields, validation...), but it works as a first
approach.

Added new [ch] files for the details dialog

Sorted lines in src/Makefile.am