frogr_facade_upload_picture to receive one picture and a callback func/object
This should help a lot to better control workflow while uploading
several pictures, while still keeping things going sequentially, thanks
to the new CPS introduced to decouple the facade from the controller by
specifying a continuation callback and object to be executed after a
picture was uploaded.

With this change, iteration over the full list of pictures is done in
the controller and the facade focuses in simple tasks only and in
executing a given callback once it's done.