Made FrogrMainWindow an abstract class, toolkit independent
* src/frogr-main-window.c: Converted into an abstract class to allow
diferent instances of FrogrMainWindow, depending on the toolkit used.
Removed a lot of non-generic code and delegate on virtual methods the
specific, toolkit dependent parts.

* src/frogr-main-window.h (frogr_main_window_new): Removed, as this
class will be an abstract class from now on.
(struct _FrogrMainWindowClass): Added needed virtual methods to
implement factory methods in the subclasses (private and public).

* src/frogr-main-window-private.h: New file to share some private
parts of FrogrMainWindow with its subclasses (kind of 'protected'
attributes and methods).

* src/Makefile.am: Added new file frogr-main-window-private.h.