More ping-pong: removed abstractions for controller and mainView
After so long since I wrote that code to test an experimental UI for
other platform (fremantle) I realized that it's very unlikely that frogr
evolves through that path so, apart from moving all the
fremantle-specific code to another branch, I decided to remove all the
hassle of having abstract controller/mainView and specific
implementations given that only one platform will be supported for now,
that is GTK (desktop) platforms, which are where frogr actually makes
sense after all.

So, it was funny to have written that code but at this moment I think
it's better to try to keep things as simple as possible and all that
stuff would only complicate things while trying to improve frogr.