Bundled the flicksoup library into frogr
This is to avoid a chicken-egg problem evolving frogr towards this
library because of it being too simple as to release a independent
package for itself. As soon as flicksoup is mature enough it would be
released in its own package but in the meanwhile, its development would
be leaded by frogr.