Simplified flicksoup (removed FspPhotosMgr) and update frogr.
In order to keep usage of fliksoup as simple as possible, I just
removed the separation between the FspSession and the FspPhotosMgr,
merging them into FspSession, so now apps using this library would
just have to create a session object and will have all the
functionality available through it.

Update frogr controller to call the right functions.