Getting rid of gtkbuilder for details dialog
Current version of Glade is behaving weirdly so I just decide I'd
go for manually coding the UI not to block some features I have in
my mind right now because of an external tool.

But for the future, it might be good to go back to gtkbuilder XML
files in order not to maintain this UI code.