Use libsoup-gnome instead of plain libsoup when available.
For the MacOSX platform plain libsoup will be always used anyway,
since it's not the mainly supported platform but an additional one,
and the main interest of using libsoup-gnome is precisely to better
integrate frogr with the GNOME platform.

Also, for cases when even in the GNOME platform one could prefer not
to use libsoup-gnome, even if available, a new parameter was added to
the configure script to specify that (--with-libsoup=plain|gnome).