Optimized usage of memory handling pixbufs internally.
So far (well, since the last commit), frogr loaded pixbufs from disk at
the maximum resolution to be used (200x200) and then stored it to be
used it both in the details window and the icon view (scaled down). This
caused a terrible waste of memory because the icon view shows quite
smaller thumbnails (140x140) and that meant that we would be storing
in memory pixbufs bigger (60x60 bigger) than what we currently needed
for most of the time just in case we wanted to display a picture in the
details window *individually*.

This patch makes frogr to store in memory the smallest pixbuf possible
(140x140, for the icon view) and retrieves the bigger pixbufs needed for
the details dialog in an on-demmand basis, saving quite a lot of memory.

Some informal tests with and without this patch shows that, for 640
high-res JPG pictures sizing 2-3MB each:

  - Without this patch: used ~120MB of memory (after finished loading)
  - With this patch: used ~60MB of memory

So the improvement is quite big, I'd say