Implemented 1st phase of authorization process in terms of OAuth.
It's now possible to make frogr open the new authorization url that
leads the user to getting a verifier code to enter later on in
frogr. Still, a lot of work is still needed, like allowing the user
introduce that verifier code somewhere to continue with the process.

But it's a good start already.