Fix problem with ref counting for instances of FrogrPhotoSet and FrogrGroup
The data type in the tree models was incorrectly declared as G_TYPE_POINTER,
instead of G_TYPE_OBJECT, causing issues when adding/removing elements to the
model (pictures not reffed/unreffed). Additionally, there were some missing
and unref calls that were causing leaks each time a dialog to add a picture
to a photoset (or create a new one) or a group was opened and then closed.

https://bugzilla.gnome.org/show_bug.cgi?id=720156