Fix cancellation of image upload process
Make sure _complete_picture_upload() is called when a picture upload
is cancelled, which will take care of stopping the process cleanly,
freeing the memory and changing the state accordingly, so that the
UI gets updated accordingly.

Also, get rid of the weird is_cancelled boolean and the callback
for the cancellable, that's not needed anymore.