Lowered gettext minimum version required down to 0.19.7
I mistakenly assumed that the new stuff needed for gettext to replace
intltool completely was added with gettext 0.19.8, and made the 1.1
release of frogr based on that wrong fact, making it complicated for
distros that are stuck in 0.19.7 to ship frogr, which is not cool.

Thus, after testing locally a build of frogr done using gettext 0.19.7
and checking that everything works ok when it comes to translations,
I'm pushing this commit now to make an quick 1.2 release on top.