Update pointer to the stable flatpak to fix the missing desktop icon
The previous version has broken icons for some languages due to a
wrongly added translation for the icon name, which made it to the
desktop file. This new commit points to a state where that issue
has been fixed, so let's update the stable flatpak while we don't
have a new release of frogr to at least fix that now.