meson: Simplify rule to install UI files
I misread the meson documentation and understood that $datadir/frogr
would get deleted if using install_subdirs() for this, but it turns
out that it would be $datadir/frogr/gtkbuilder instead, and so there
is no problem on using install_subdirs().

Therefore, use that rule instead and simplify these lines.