Implement workaround to load images when GdkPixbufLoader fails to do so
There seems to be an obscure bug in gdk-pixbuf that causes to fail loading
images into GdkPixbuf instances when using GdkPixbufLoader, even though
the same images load just fine with gdk_pixbuf_new_from_file().

I've investigated this a bit in gdk-pixbuf but couldn't find the problem
myself so, considering that this bug has been reported a while ago already
and that it's very annoying (specially when uploading certain pictures taken
with mobile phones), I've decided to push this workaround for frogr hoping
that this will get fixed in gdk-pixbuf, which I'll file a bug about it.

https://bugzilla.gnome.org/show_bug.cgi?id=768639