meson: Stop including the configure wrapper script in the dist tarball
Shipping this script with the tarball is causing confusiong and trouble
for some distros when packaging (e.g. Debian), since they see that file
and assume it's a regular autotools project, which is not the case.

As this has been added for compatibility with GNOME Continuous, let's
stop including it in the distribution tarball to force packagers to
use meson natively.