Remove the application menu (while preserving the same functionality)
This is in line with the "Application Menu Retirement" initiative [1],
aiming to reduce usablitiy issues since the AppMenu was introduced.

In builds with the header menu enabled, this means that the contents
of the AppMenu have been moved to the "Hamburger" menu on the right
side of the header bar. For builds configured without a header bar,
the items will be inside the "File" menu of the traditional menu bar.

[1] https://gitlab.gnome.org/GNOME/Initiatives/wikis/App-Menu-Retirement