Set Werror=false by default
This makes sense for dev builds but can potentially break the build
in environment where any library frogr depends on is building with
warnings, as that would of course break the build.

This has been recently the case for Hurd and FreeBSD (detected by
Debian), so let's prevent thi from happening by setting saner defaults.