Bring back the possibility of building with libsoup 2.4
Now that everything has been migrated to libsoup3, let's re-add
the possibility of building with libsoup2 to allow building against
the older API, at least for a while.

In order to do that, just pass -Dwith-libsoup2=true to meson setup
when configuring the build, which will add a WITH_LIBSOUP2 define
to the config.h file if that is the case, and then use that to
if-def the different codepaths to build with libsoup2 or libsoup3.